check if exist previous_field_location.log

+++ sendIM/sendBMEDataToIM_periodically.py
@@ -361,18 +361,22 @@ logger.info('電源モジュールに送信するコマンド用意：' + powerC
 #もろもろ書かれたlogは結局boot.logに上書きされているので、
 #検討するのはboot.logのみでよい
 
-file_name = "previous_field_location.log"
-try:
-    _timeStamp = sendLog_ftps(file_name, put_directory)
 
-    #log送信正常終了なので、中身をクリアする
-    with codecs.open('/var/log/' + file_name, 'w', 'utf_8_sig') as f:
-#            f.write(unicode(codecs.BOM_UTF8, 'utf_8'))
-        f.write (u'アップロード終了 with no error. Log cleared at: ' + _timeStamp.strftime('%Y%m%d%H%M') + '\n'.encode('utf_8'))
-    f.close()
-except Exception as e:
-        logger.debug("sendLog_ftps error. :" + str(e))
+file_name = "previous_field_location.log"
+
+src = '/var/log/' + file_name
+if os.path.isfile(src):
+    try:
+        _timeStamp = sendLog_ftps(file_name, put_directory)
+
+        #log送信正常終了なので、中身をクリアする
+        with codecs.open('/var/log/' + file_name, 'w', 'utf_8_sig') as f:
+    #            f.write(unicode(codecs.BOM_UTF8, 'utf_8'))
+            f.write (u'アップロード終了 with no error. Log cleared at: ' + _timeStamp.strftime('%Y%m%d%H%M') + '\n'.encode('utf_8'))
         f.close()
+    except Exception as e:
+            logger.debug("sendLog_ftps error. :" + str(e))
+            f.close()
 
 file_name = "boot.log"
 try:
